chore: make the code and files consistent with the current repo
