+++ application/slack/v0/task_functions.go
@@ -10,6 +10,42 @@ import (
 	"google.golang.org/protobuf/types/known/structpb"
 )
 
+type UserInputReadTask struct {
+	ChannelName     string `json:"channel_name"`
+	StartToReadDate string `json:"start_to_read_date"`
+	IsPublicChannel bool   `json:"is_public_channel"`
+}
+
+type ReadTaskResp struct {
+	Conversations []Conversation `json:"conversations"`
+}
+
+type Conversation struct {
+	UserID             string               `json:"user_id"`
+	Message            string               `json:"message"`
+	StartDate          string               `json:"start_date"`
+	LastDate           string               `json:"last_date"`
+	TS                 string               `json:"ts"`
+	ReplyCount         int                  `json:"reply_count"`
+	ThreadReplyMessage []ThreadReplyMessage `json:"thread_reply_messages"`
+}
+
+type ThreadReplyMessage struct {
+	UserID   string `json:"user_id"`
+	DateTime string `json:"datetime"`
+	Message  string `json:"message"`
+}
+
+type UserInputWriteTask struct {
+	ChannelName     string `json:"channel_name"`
+	Message         string `json:"message"`
+	IsPublicChannel bool   `json:"is_public_channel"`
+}
+
+type WriteTaskResp struct {
+	Result string `json:"result"`
+}
+
 func (e *execution) readMessage(in *structpb.Struct) (*structpb.Struct, error) {
 
 	params := UserInputReadTask{}
